+++ backend/api/serializers.py
@@ -16,7 +16,7 @@ from rest_framework import serializers, validators
 from organizations.models import (Appointment, District,
                                   Organization, OrganizationSpecialty,
                                   OrganizationBusinessHour,
-                                  Specialty, Town, Review)
+                                  Specialty, Town)
 from user.models import User
 from .fields import DistrictField, SlugRelatedFieldWith404
 from .utils import FIO_REGEX, PHONE_NUMBER_REGEX, haversine
@@ -525,30 +525,3 @@ class PasswordResetConfirmRetypeSerializer(
     UidAndTokenSerializer переопределен.
     """
     pass
-
-
-class ReviewSerializer(serializers.ModelSerializer):
-    author = serializers.SlugRelatedField(
-        slug_field='email',
-        read_only=True,
-    )
-
-    score = serializers.IntegerField(min_value=1, max_value=5)
-
-    class Meta:
-        model = Review
-        fields = ('id', 'text', 'author', 'score', 'pub_date')
-        read_only_fields = ('id',)
-
-    def validate(self, data):
-        request = self.context.get('request')
-
-        if request.method == 'POST':
-            uuid = self.context['view'].kwargs.get('uuid')
-            uuid_name = get_object_or_404(Organization, uuid=uuid)
-            if Review.objects.filter(
-                    author=request.user,
-                    organization=uuid_name
-            ).exists():
-                raise serializers.ValidationError('Вы уже оставили отзыв!')
-        return data
