+++ backend/api/views/__init__.py
@@ -3,4 +3,3 @@ from .organization import OrganizationViewSet
 from .specialty import SpecialtyViewSet
 from .town import TownViewSet
 from .auth import SignUp, LoginView
-from .review import ReviewViewSet
